+++ docs/conf.py
@@ -42,7 +42,10 @@ extensions = [
     "sphinx.ext.viewcode",
     "sphinx_toolbox",
     "sphinx_toolbox.github",
+    "sphinx_toolbox.more_autodoc.autoprotocol",
+    "sphinx_toolbox.more_autodoc.sourcelink",
     "sphinx_toolbox.sidebar_links",
+    "sphinx_toolbox.wikipedia",
 ]
 exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 html_theme = "furo"
@@ -60,6 +63,9 @@ napoleon_attr_annotations = True
 html_permalinks_icon = "<span>#</span>"
 github_username = author
 github_repository = project
+autodoc_show_sourcelink = True
+todo_include_todos = True
+todo_emit_warnings = True
 
 
 def badge_flstudio(app, what, name, obj, options, lines):